% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{pretty.qca}
\alias{pretty.qca}
\title{Print Pretty QCA Solution}
\usage{
\method{pretty}{qca}(sol, number = 1)
}
\arguments{
\item{sol}{a qca object}

\item{number}{which intermediate solution}
}
\description{
Prints a pretty version of the QCA solution that can be used with knitr and rmarkdown
}
