% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{pretty.tt}
\alias{pretty.tt}
\title{Print Pretty Truth Table}
\usage{
\method{pretty}{tt}(truth, complete = FALSE)
}
\arguments{
\item{truth}{a truthTable object}

\item{complete}{print the complete truth table, including empty rows?}
}
\value{
A markdown formatted version of the truth table
}
\description{
prints a pretty version of the truth table for use with knitr and rmarkdwon
}
